Move/cleanup Flusher.
